feat(email): Improving email templates
This improves how the email templates are built, as well as consolidates
some parts of the template into a re-usable component. It also adds the
trail expiry email which will be used soon in a follow up PR.
