+++ emails/CMakeLists.txt
@@ -1,4 +1,4 @@
-set(EMAIL_TEMPLATE_INPUT_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}/src")
+set(EMAIL_TEMPLATE_INPUT_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}/src/emails")
 set(EMAIL_TEMPLATE_HTML_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/emails/output/html")
 set(EMAIL_TEMPLATE_TXT_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/emails/output/html")
 file(MAKE_DIRECTORY ${EMAIL_TEMPLATE_HTML_OUTPUT_DIRECTORY})
@@ -20,7 +20,7 @@ list(TRANSFORM TEXT_EMAIL_TEMPLATES REPLACE ".tsx" ".txt")
 add_custom_command(
   OUTPUT ${HTML_EMAIL_TEMPLATES}
   BYPRODUCTS ${HTML_EMAIL_TEMPLATES}
-  COMMAND ${REACT_EMAIL_EXECUTABLE} export --dir src --outDir ${EMAIL_TEMPLATE_HTML_OUTPUT_DIRECTORY}
+  COMMAND ${REACT_EMAIL_EXECUTABLE} export --dir src/emails --outDir ${EMAIL_TEMPLATE_HTML_OUTPUT_DIRECTORY}
   COMMAND ${CMAKE_COMMAND} -E copy ${EMAIL_TEMPLATE_HTML_OUTPUT_DIRECTORY}/* ${EMAIL_IN_TREE_OUTPUT}
   COMMENT "Building HTML email templates from React"
   WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
@@ -32,7 +32,7 @@ add_custom_command(
 add_custom_command(
   OUTPUT ${TEXT_EMAIL_TEMPLATES}
   BYPRODUCTS ${TEXT_EMAIL_TEMPLATES}
-  COMMAND ${REACT_EMAIL_EXECUTABLE} export --plainText --dir src --outDir ${EMAIL_TEMPLATE_TXT_OUTPUT_DIRECTORY}
+  COMMAND ${REACT_EMAIL_EXECUTABLE} export --plainText --dir src/emails --outDir ${EMAIL_TEMPLATE_TXT_OUTPUT_DIRECTORY}
   COMMAND ${CMAKE_COMMAND} -E copy ${EMAIL_TEMPLATE_TXT_OUTPUT_DIRECTORY}/* ${EMAIL_IN_TREE_OUTPUT}
   COMMENT "Building TXT email templates from React"
   WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
@@ -52,7 +52,7 @@ add_custom_target(
 add_custom_target(
   development.email
   COMMENT "Running the local development tools for email templates"
-  COMMAND ${REACT_EMAIL_EXECUTABLE} dev --dir src
+  COMMAND ${REACT_EMAIL_EXECUTABLE} dev --dir src/emails
   WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
   DEPENDS
     tools.react-email
